backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
  "flask>=3.0.0",
  "flask-cors>=4.0.0",
  "python-dotenv>=1.0.0",
  "pydantic>=2.12.0",
  "livekit-api>=1.1.0",
  "livekit-agents[google]~=1.4",
  "livekit-plugins-deepgram",
  "livekit-plugins-google",
  "livekit-plugins-openai",
  "livekit-plugins-spatialreal",
  "langgraph>=0.2.0",
  "langchain-openai>=0.2.0",
  "langchain-core>=0.3.0",
  "livekit-plugins-silero",
  "livekit-plugins-ai-coustics",
  "livekit-plugins-turn-detector",
  "psycopg2-binary>=2.9.11",
]
